+++ Sources/PatchDay/Sites/SitesViewModel.swift
@@ -94,7 +94,6 @@ class SitesViewModel: CodeBehindDependencies<SitesViewModel>, SitesViewModelProt
         guard let sites = sdk?.sites else { return }
         sites.delete(at: indexPath.row)
         table.deleteCell(indexPath: indexPath)
-        table.reloadCells()
     }
 
     func createBarItems(